Final update
I added some buildings and textures them to make a makeshift village without being having to fully model and lay out the entirety of the buildings. I added fire and screaming effects to one of the houses further away from the main cluster to act at the shaman's house that is on fire as the shaman is being burnt alive.